# toc-backend/api/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: toc
  description: An API to create backup and restoration
  termsOfService: "http://gyk4j.github.io/terms/"
  contact:
    email: "147011991+gyk4j@users.noreply.github.com"
  license:
    name: "Eclipse Public License 2.0 (EPL)"
    url: "https://www.eclipse.org/legal/epl-2.0/"
host: 0.0.0.0:80
basePath: /v1
...
schemes:
  - "http"
definitions:
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        x-omitempty: false
      type:
        type: "string"
        x-omitempty: false
        description: "Severity level"
        enum: [ error, warning, info, debug, trace ]
      message:
        type: "string"
        x-omitempty: false
    example:
      code: 0
      type: "info"
      message: "message"
  Snapshot:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        x-omitempty: false
      file:
        type: "string"
        x-omitempty: false
      status:
        type: "string"
        x-omitempty: false
        description: "Snapshot Status"
        enum: [ queued, in-progress, completed, failed ]
      complete:
        type: "boolean"
        x-omitempty: false
    xml:
      name: "snapshot"
    example:
      file: "/var/opt/toc/archive/20241231T235959Z-server.tar.gz"
      id: 1
      status: "completed"
      complete: true
  Backup:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        x-omitempty: false
      time:
        type: "string"
        format: "date-time"
        x-omitempty: false
      snapshots:
        type: "array"
        x-omitempty: false
        items:
          $ref: "#/definitions/Snapshot"
    xml:
      name: "Backup"
    example:
      time: "2024-12-31T23:59:59Z"
      id: 1
...
  Transfer:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        x-omitempty: false
      backup:
        $ref: "#/definitions/Backup"
        x-omitempty: false
      status:
        type: "string"
        x-omitempty: false
        description: "Transfer Status"
        enum: [ queued, in-progress, completed, failed ]
    xml:
      name: "Transfer"
    example:
      id: 1
      status: "failed"
...
paths:
...
  /transfers:
    post:
      tags:
        - "transfer"
      summary: "Start a backup transfer attempt"
      description: "Run transfer and return immediately without waiting"
      operationId: "newTransfer"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Backup and associated snapshots for restoration"
          required: true
          schema:
            $ref: "#/definitions/Backup"
          x-exportParamName: "Body"
      responses:
        "200":
          description: "Transfer started"
          schema:
            $ref: "#/definitions/Transfer"
        "403":
          description: "Forbidden from creating duplicate/repeat transfer"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "Stepup backup server not found"
          schema:
            $ref: "#/definitions/ApiResponse"
        "405":
          description: "Bad request"
          schema:
            $ref: "#/definitions/ApiResponse"
        "500":
          description: "Stepup TOC controller error"
          schema:
            $ref: "#/definitions/ApiResponse"
        "502":
          description: "Bad gateway. Stepup backup server error"
          schema:
            $ref: "#/definitions/ApiResponse"
        "503":
          description: "Service unavailable on stepup TOC controller"
          schema:
            $ref: "#/definitions/ApiResponse"
        "504":
          description: "Gateway timeout. Stepup backup server did not reply."
          schema:
            $ref: "#/definitions/ApiResponse"
    put:
      tags:
        - "transfer"
      summary: "Update transfer status/info"
      description: "Update transfer status/info"
      operationId: "updateTransfer"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Transfer request to update"
          required: true
          schema:
            $ref: "#/definitions/Transfer"
          x-exportParamName: "Body"
      responses:
        "200":
          description: "Transfer updated"
          schema:
            $ref: "#/definitions/Transfer"
        "403":
          description: "Forbidden from updating transfer"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "TOC controller not found"
          schema:
            $ref: "#/definitions/ApiResponse"
        "405":
          description: "Bad request"
          schema:
            $ref: "#/definitions/ApiResponse"
        "500":
          description: "TOC controller error"
          schema:
            $ref: "#/definitions/ApiResponse"
        "502":
          description: "Bad gateway. TOC controller error"
          schema:
            $ref: "#/definitions/ApiResponse"
        "503":
          description: "Service unavailable on TOC controller"
          schema:
            $ref: "#/definitions/ApiResponse"
        "504":
          description: "Gateway timeout. TOC controller did not reply."
          schema:
            $ref: "#/definitions/ApiResponse"
    get:
      tags:
        - "transfer"
      summary: "Query transfer status"
      description: "Retrieve status of all transfers"
      operationId: "getTransfers"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Transfer info retrieved"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Transfer"
        "403":
          description: "Forbidden from querying transfer status"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "Transfer not found"
          schema:
            $ref: "#/definitions/ApiResponse"
        "405":
          description: "Bad request"
          schema:
            $ref: "#/definitions/ApiResponse"
        "500":
          description: "TOC controller error"
          schema:
            $ref: "#/definitions/ApiResponse"
        "503":
          description: "Service unavailable"
          schema:
            $ref: "#/definitions/ApiResponse"
  /transfers/{transferId}:
    get:
      tags:
        - "transfer"
      summary: "Find transfer by ID"
      description: "Query a single transfer"
      operationId: "getTransferById"
      produces:
        - "application/json"
      parameters:
        - name: "transferId"
          in: "path"
          description: "ID of transfer to return"
          required: true
          type: "integer"
          format: "int64"
          x-exportParamName: "TransferId"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Transfer"
        "400":
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "Transfer not found"
          schema:
            $ref: "#/definitions/ApiResponse"
...